Address a comment + move MeshBuffers
